move Prop to a separate file
